Extract BitchatPacket into a separate file
